WIP Update Code for Azure and GCP * Split code into 4 files * Clean up many methods * Start translating s3 methods to az methods * Change names of some global vars and env vars
